Remove IoCoroutineWorker; use SyncDispatcher in BaseSyncWorker and PushRegistrationWorker
