+++ core/src/main/kotlin/at/bitfire/davdroid/push/PushRegistrationWorker.kt
@@ -6,10 +6,13 @@ package at.bitfire.davdroid.push
 
 import android.content.Context
 import androidx.hilt.work.HiltWorker
+import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
-import at.bitfire.davdroid.IoCoroutineWorker
+import at.bitfire.davdroid.di.qualifier.SyncDispatcher
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.withContext
 import java.util.logging.Logger
 
 /**
@@ -23,10 +26,16 @@ class PushRegistrationWorker @AssistedInject constructor(
     @Assisted context: Context,
     @Assisted workerParameters: WorkerParameters,
     private val logger: Logger,
-    private val pushRegistrationManager: PushRegistrationManager
-) : IoCoroutineWorker(context, workerParameters) {
+    private val pushRegistrationManager: PushRegistrationManager,
+    @SyncDispatcher private val syncDispatcher: CoroutineDispatcher
+) : CoroutineWorker(context, workerParameters) {
 
-    override suspend fun doIoWork(): Result {
+    override suspend fun doWork(): Result =
+        withContext(syncDispatcher) {   // required because of Ktor issue; see SyncDispatcher KDoc
+            updatePushRegistrations()
+        }
+
+    suspend fun updatePushRegistrations(): Result {
         logger.info("Running push registration worker")
 
         // update registrations for all services
